integration: move test_shard_out_of_range to shards.rs
`test_shard_out_of_range()` was put in `load_balancing.rs`, while
`shards.rs` seems to be a much better place for it.
